refactor(screenscraper): extract media-resolution helpers into a collaborator
Move the HTTP-free media-resolution cluster out of ScreenScraperService into a
new services/screenscraper/media_resolver.dart collaborator:

- getMediaDirectory    — resolves/caches the media dir (was _getMediaDirectory)
- mapMediaTypeToFolder  — API type -> folder name (was _mapMediaTypeToFolder)
- selectBestMedia       — best regional/language asset (was _selectBestMedia)
- checkFileExists       — local asset existence check (was _checkFileExists)

The collaborator becomes the single owner of the media-directory cache
(_cachedMediaDirectory moved with getMediaDirectory). None of these perform
network I/O, so the move is verbatim (token-stream byte-identical vs main); only
delta is the private->public rename. Callers now delegate to
ScreenscraperMediaResolver. Dropped the now host-unused config_service import.
The HTTP downloaders (_downloadGameMedia/_downloadMediaFileSmart) stay on the
facade until the client collaborator is extracted. No public-API change — the 11
facade methods are untouched.
